Time: 23 ms (67.60%), Space: 22.4 MB (92.56%) - LeetHub

+++ 0637-average-of-levels-in-binary-tree/0637-average-of-levels-in-binary-tree.cpp
@@ -19,7 +19,7 @@ public:
         q.push(root);
         for (int level = 0, sz = q.size(); !q.empty(); level++, sz = q.size())
         {
-            double size = sz;
+            int size = sz;
             double sum = 0;
             while (sz--)
             {
